requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
clarifai==2.6.2
clarifai-grpc==9.0.0
configparser==3.8.1
dnspython==2.3.0
ffmpeg==1.4
ffprobe==0.5
ffprobe-python==1.0.3
future==0.18.3
googleapis-common-protos==1.58.0
grpcio==1.51.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jsonschema==2.6.0
protobuf==3.20.3
pydub==0.25.1
pymongo==4.3.3
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
urllib3==1.26.14
